+++ foxy_farmer/foxy_farmer_main.py
@@ -1,3 +1,6 @@
+import sys
+from os import execv, spawnv, P_WAIT, _exit
+
 import click
 from multiprocessing import freeze_support
 from pathlib import Path
@@ -16,16 +19,18 @@ from foxy_farmer.util.root_path import get_root_path
 from foxy_farmer.version import version
 
 
-async def run_foxy_farmer(foxy_root: Path, config_path: Path):
+async def run_foxy_farmer(foxy_root: Path, config_path: Path) -> bool:
     from foxy_farmer.foxy_farmer_class import FoxyFarmer
     foxy_farmer = FoxyFarmer(foxy_root, config_path)
     await foxy_farmer.setup_process_global_state()
     try:
-        await foxy_farmer.run()
+        return await foxy_farmer.run()
     except AlreadyRunningException:
         pass
     finally:
         close_sentry()
+
+    return False
 
 
 @click.group(
@@ -63,7 +68,14 @@ def cli(ctx, config, root_path):
 @click.pass_context
 def run_cmd(ctx, config=None, root_path=None):
     from chia.server.start_service import async_run
-    async_run(run_foxy_farmer(ctx.obj["root_path"], ctx.obj["config_path"]))
+    should_restart_app = async_run(run_foxy_farmer(ctx.obj["root_path"], ctx.obj["config_path"]))
+
+    if should_restart_app:
+        if sys.platform == "win32":
+            # On Windows we can not replace the process, spawn subprocess and wait for its termination instead
+            _exit(spawnv(P_WAIT, sys.executable, sys.argv))
+        else:
+            execv(sys.executable, sys.argv)
 
 
 cli.add_command(summary_cmd)