+++ foxy_farmer/util/is_binary.py
@@ -1,0 +1,5 @@
+import sys
+
+
+def is_binary() -> bool:
+    return getattr(sys, 'frozen', False)